aib/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
anthropic>=0.39.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
PyPDF2>=3.0.1
Pillow>=10.4.0
slowapi>=0.1.9
python-multipart>=0.0.12
